refactor: 장소 DB 처리 코드 개선
- local, remote DB 객체 연결
