+++ app/src/main/java/campus/tech/kakao/map/PlaceApplication.kt
@@ -4,13 +4,16 @@ import android.app.Application
 import android.net.ConnectivityManager
 import android.net.Network
 import android.net.NetworkCapabilities
+import campus.tech.kakao.map.data.PlaceDBHelper
 import campus.tech.kakao.map.data.PlaceRepositoryImpl
 import campus.tech.kakao.map.domain.repository.PlaceRepository
 import com.kakao.vectormap.KakaoMapSdk
 
 class PlaceApplication: Application() {
 
-    val placeRepository: PlaceRepository by lazy { PlaceRepositoryImpl.getInstance(this)}
+    val placeRepository: PlaceRepository by lazy {
+        PlaceRepositoryImpl(PlaceDBHelper.getInstance(this))
+    }
 
     override fun onCreate() {
         super.onCreate()
